+++ lib/src/utf8.dart
@@ -35,7 +35,9 @@ Uint8List toUtf8(
 @pragma('vm:prefer-inline')
 @pragma('dart2js:tryInline')
 String fromUtf8(
-  Uint8List input, {
+  List<int> input, {
   UTF8Codec codec = UTF8Codec.standard,
 }) =>
-    codec.decoder.decode(input);
+    codec.decoder.decode(
+      input is Uint8List ? input : Uint8List.fromList(input),
+    );